requirements.txt
redis
pyodbc
psutil
python-dotenv
